Bug 4610: compile errors on Solaris 11.3 with Oracle Studio 12.5
Note: this patch is incomplete. Additional changes are required 
to fully build Squid-4. see the bug report for details.

+++ src/HttpHdrCc.cc
@@ -306,7 +306,7 @@ httpHdrCcStatDumper(StoreEntry * sentry, int, double val, double, int count)
 {
     extern const HttpHeaderStat *dump_stat; /* argh! */
     const int id = static_cast<int>(val);
-    const bool valid_id = id >= 0 && id < HttpHdrCcType::CC_ENUM_END;
+    const bool valid_id = id >= 0 && id < static_cast<int>(HttpHdrCcType::CC_ENUM_END);
     const char *name = valid_id ? CcAttrs[id].name : "INVALID";
 
     if (count || valid_id)
